# packages/contracts/openapi/depsis.yaml
openapi: 3.1.0
info:
  title: DEPSIS API
  version: 0.1.0-phase1
  summary: Debian tabanlı çok kullanıcılı NAS platformu — Faz 1 yüzeyi
  description: |
    Bu belge tek doğruluk kaynağıdır. `apps/web` istemcisi buradan üretilir
    (`openapi-typescript`, ADR-0002); elle yazılmış istemci tipi bulunamaz.

    Kapsam: Faz 1 uçları. Konsol, uygulama kataloğu ve uzak erişim de burada — sırasıyla ADR-0018, ADR-0019 ve ADR-0020. Üçü de cihazın DIŞINDAKİ bir şeyi yönetiyor (bir kabuk, podman, zerotier-one); DEPSIS hiçbirini paketlemez, kuruluysa yönetir ve kurulu değilse 503 döner. Konsol ayrıcalıklı ajanda ÇALIŞMAZ: §2.2'nin serbest komut yasağı duruyor.
    Faz 2-3'e aittir ve kasıtlı olarak yoktur — tanımlanmamış bir uç, "yakında" diye
    duran boş bir uçtan iyidir.

    Sistem aracısı IPC şeması burada DEĞİLDİR. ADR-0006 gereği onu Rust sahiplenir ve
    schemars ile üretir: bir güven sınırının sözleşmesini, o sınırı uygulayan taraf
    tanımlamalıdır.
servers:
  - url: /api/v1
...
security:
  - sessionCookie: []
paths:
...
  /setup/claim:
    post:
      tags: [setup]
      summary: Sistemi tek seferlik belirteçle sahiplen
      security: []
      description: |
        İlk organizasyonu ve yöneticisini yaratır. Belirteç, kurulum beklerken API'nin her
        açılışta journal'a bastığı tek kullanımlık değerdir; okumak konsol veya SSH erişimi
        gerektirir ve ilk yöneticinin kim olacağına karar vermesi gereken yetki budur
        (ADR-0015 §5d).

        Doğrulama hataları burada ALANI SÖYLER, giriş yolunun aksine: karşı taraf makinenin
        sahibi, formu bir kez dolduruyor.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetupClaimRequest' }
      responses:
        '200':
          description: Sahiplenildi
          content:
            application/json:
              schema:
                type: object
                required: [status, organizationSlug]
                properties:
                  status: { type: string, enum: [ok] }
                  organizationSlug: { type: string }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '410':
          description: Kurulum zaten tamamlandı; bu uç kalıcı olarak kapalı
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
...
  /shares:
    get:
      tags: [system]
      summary: Paylaşımlar ve SMB adresleri
      description: |
        §9. Bir NAS'ın varlık sebebi Windows'tan bir sürücü olarak görünmek, ve o adresi
        kullanıcıya SÖYLEYEN bir yer yoksa özellik yarım kalır — kimse `\\depsis\belgeler`
        yazacağını tahmin edemez.

        `published`, ajanın Samba yapılandırmasını gerçekten yazıp yazmadığını söyler. Bu alan
        iyimser DEĞİL: DEPSIS'in veritabanında bir paylaşım satırı olması, smbd'nin onu sunduğu
        anlamına gelmiyor, ve ikisini aynı şeymiş gibi göstermek kullanıcıyı çalışmayan bir
        adresi denemeye gönderir.
      responses:
        '200':
          description: Paylaşımlar
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SharePage' }

    post:
      tags: [system]
      summary: Paylaşım aç
      description: |
        Yalnız yöneticiler. §9'un eksik yarısı: bugüne kadar DEPSIS paylaşımları LİSTELİYOR ama
        açamıyordu — satırlar yalnız testlerden ve elle gelebiliyordu.

        Üç şey tek bir işlemde olur ve sırası önemlidir:

        1. **Dataset.** Ajan `zfs create -o acltype=posixacl` çalıştırır ve özelliği geri OKUR;
           `nfsv4` kendini kurulmuş gibi raporlayıp hiçbir şey uygulamadığı için (P0-B) bu
           doğrulama pazarlık konusu değil. Dataset `DEPSIS_SHARE_PARENT_DATASET` altına açılır
           ve oraya bağlı olduğu için paylaşım köküne düşer.
        2. **Satır.** `public.shares`. Dataset ÖNCE açılır: bir satır yazıp dataset'i açamamak,
           dosya yöneticisinde görünen ama açılamayan bir paylaşım bırakır.
        3. **Kök grant.** `grants` verilmezse paylaşımı açan yöneticiye tam izin yazılır.

        Üçüncüsü isteğe bağlı DEĞİL, ve sebebi bir güvenlik açığı: grant'ı olmayan bir paylaşım,
        §6.2 öncesi "kiracının herkesi yedi izin alır" davranışına geri düşen bir istisnayı
        canlı tutuyordu. `POST /shares` kök grant'ı aynı işlemde yazdığı için o istisna
        kaldırıldı (migration 0016). **Yeni paylaşım KAPALI doğar**: yalnız grant'ta adı geçen
        principal'lar ve organizasyon yöneticileri görür.

        POSIX ACL'lerin yazılması bir iş kuyruğuna girer ve `applyingJobId` onu adlandırır.
        `null` ise ajana ulaşılamadı: satır ve grant duruyor, dosya sistemi henüz haberdar değil.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ShareCreateRequest' }
      responses:
        '201':
          description: Açıldı
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ShareCreated' }
        '403': { $ref: '#/components/responses/Problem' }
        '409':
          # Ad zaten alınmış (veritabanında ya da dataset olarak).
          $ref: '#/components/responses/Problem'
        '422': { $ref: '#/components/responses/Problem' }
        '503':
          # Ajana ulaşılamıyor ya da DEPSIS_SHARE_PARENT_DATASET ayarlı değil.
          $ref: '#/components/responses/Problem'
...
  /backups:
    get:
      tags: [system]
      summary: DEPSIS'in aldığı anlık görüntüler
      description: |
        Havuzdaki anlık görüntülerin listesi DEĞİL, ve bu fark önemli. Ajanın işlem kümesi kapalı
        (ADR-0006) ve içinde `CreateSnapshot` ile `DiffSnapshots` var ama "listele" yok. Biri
        Rust tarafına o işlemi ekleyene kadar dürüstçe gösterilebilecek şey, DEPSIS'in kendi
        aldıklarının kaydıdır. Kabuktan alınmış bir anlık görüntü burada görünmez ve arayüz bunu
        söylemek zorundadır.
      responses:
        '200':
          description: Anlık görüntüler
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SnapshotPage' }
        '403': { $ref: '#/components/responses/Problem' }
    post:
      tags: [system]
      summary: Anlık görüntü al
      description: |
        Yalnız yöneticiler. Ajanın `CreateSnapshot` işlemi çağrılır; ajan doğrularsa satır yazılır.
        Sıra bu yönde: önce satır yazıp sonra ajanı çağırmak, ajan reddettiğinde var olmayan bir
        yedeği listeleyen bir kayıt bırakırdı.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSnapshotRequest' }
      responses:
        '201':
          description: Alındı
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Snapshot' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
        '503': { $ref: '#/components/responses/Problem' }
...
  /files/{id}/permissions:
    parameters:
      - $ref: '#/components/parameters/FileId'
    get:
      tags: [permissions]
      summary: Bu klasörün izinleri
      description: |
        İki ayrı şey döner ve karıştırılmamalı:

        - `effective` — ÇAĞIRANIN bu klasörde gerçekten yapabildikleri, miras hesaplandıktan
          sonra. Arayüz düğmeleri buna göre çizer.
        - `grants` — bu düğümde AÇIKÇA yazılmış izin satırları. Yalnız `manage` iznine sahip
          olana döner; olmayan için boş dizi gelir, çünkü kimin nereye erişebildiği kendi başına
          bir bilgidir.

        `inheritedFrom`, `effective` kümesinin hangi atadan geldiğini söyler (bu düğümün kendisi
        ise kendi kimliği, hiç grant yoksa null). ADR-0021'in miras kuralı temsilci başına EN
        YAKIN atayı seçtiği için "bu izin nereden geliyor" cevaplanabilir bir soru ve
        cevaplanmalı — aksi hâlde bir yöneticinin izni nereden kaldıracağını bulması imkânsız.
      responses:
        '200':
          description: İzinler
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FolderPermissions' }
        '404': { $ref: '#/components/responses/Problem' }
    put:
      tags: [permissions]
      summary: Bu klasörün açık izinlerini yaz
      description: |
        `manage` izni gerekiyor — `modify` YETMEZ. Bir klasöre yazabilen herkesin o klasörün
        iznini de değiştirebilmesi, yetki modelinin kendi kendini geçersiz kılmasıdır (ADR-0021).

        Gövde bu düğümdeki grant'ların TAMAMINI değiştirir. Kısmi birleştirme, iki yöneticinin
        birbirinin izin listesini yarım yazması demek olurdu.

        `?dryRun=true` ile hiçbir şey yazılmaz ve yalnız etki döner. §6.2 bunu şart koşuyor ve
        gerekçesi somut: miras bir alt ağacın tamamını değiştiriyor, yani "şu klasörü ekibe
        açayım" tıklaması kişinin farkında olmadığı beş yüz dosyayı açabilir.

        İzinler dosya sistemine ANINDA yansımaz. POSIX ACL'lerin yeniden yazılması bir iş
        kuyruğu görevi ve tamamlanana kadar iki katman ayrışıktır (ADR-0021'in kabul edilen
        borcu); yanıt `applyingJobId` ile bunu söyler ve arayüz "izinler uygulanıyor"
        gösterebilir.
      parameters:
        - name: dryRun
          in: query
          schema: { type: boolean, default: false }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetFolderPermissionsRequest' }
      responses:
        '200':
          description: Yazıldı, ya da dryRun ise ön izleme
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PermissionWriteResult' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: depsis_session
      description: |
        Sunucu tarafı oturum. JWT kullanılmaz — güvenlik olayında tüm oturumların iptal
        edilebilmesi gerekiyor ve stateless token bunu ucuz yapmıyor (ADR-0009).
  parameters:
    FileId:
      name: id
      in: path
      required: true
      description: Kararlı dosya kimliği. Yol asla kimlik olarak kullanılmaz (ADR-0005).
      schema: { type: string, format: uuid }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: |
        Kullanıcı ve uç kapsamında benzersiz. Aynı anahtar farklı bir istek parmak iziyle
        gelirse 409 döner. RFC yoktur — IETF taslağı Expired and archived durumundadır — bu
        yüzden semantik burada tanımlanır (ADR-0008).
      schema: { type: string, maxLength: 255 }
...
  responses:
    Problem:
      description: RFC 9457 Problem Details
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
  schemas:
    ProblemDetails:
      type: object
      required: [type, title, status, code, correlationId]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail:
          type: string
          description: Secret, dosya içeriği veya çağıranın görmeye yetkili olmadığı yol içermez.
        instance: { type: string }
        code:
          type: string
          description: Kapalı küme. Kaynak tanım packages/contracts/src/errors.ts.
        correlationId:
          type: string
          description: Her yanıtta ve aynı isteğin her log satırında bulunur (§14).
        errors:
          type: array
          items: { $ref: '#/components/schemas/ProblemFieldError' }
        retryAfter: { type: integer, minimum: 0 }
...
    ProblemFieldError:
      type: object
      required: [pointer, code, message]
      properties:
        pointer: { type: string }
        code: { type: string }
        message: { type: string }
...
    SetupClaimRequest:
      type: object
      required: [token, organizationSlug, organizationName, adminUsername, adminPassword]
      properties:
        token: { type: string, maxLength: 128 }
        organizationSlug:
          type: string
          maxLength: 63
          pattern: '^[a-z0-9]([a-z0-9-]{0,61}[a-z0-9])?$'
        organizationName: { type: string, maxLength: 200 }
        adminUsername:
          type: string
          maxLength: 64
          pattern: '^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$'
        adminPassword:
          type: string
          minLength: 12
          maxLength: 1024
          description: |
            Yalnız uzunluk tabanı. Bileşim kuralı (bir büyük, bir rakam, bir sembol) arama
            uzayını genişletmekten çok daraltıyor ve NIST SP 800-63B tam bu yüzden tavsiyeyi
            geri çekti.
...
    Snapshot:
      type: object
      required: [id, dataset, name, fullName, createdAt, state]
      properties:
        id:
          type: [string, 'null']
          format: uuid
          description: |
            DEPSIS'in kaydındaki kimlik. `state: unmanaged` olan bir satırda **null**: o görüntü
            havuzda var ama DEPSIS'in onun için bir satırı yok, ve uydurulmuş bir kimlik istemciye
            gerçek olmayan bir tutamak verirdi.
        dataset: { type: string }
        name: { type: string }
        fullName: { type: string }
        createdBy: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        state:
          type: string
          enum: [present, missing, unmanaged, unknown]
          description: |
            Kayıt ile HAVUZUN karşılaştırılması. Üçü de okuyanın yapacağı şeyi değiştiriyor:

            * `present` — kayıtta var, havuzda var. Güvenilir bir geri dönüş noktası.
            * `missing` — kayıtta var, havuzda YOK. Kabuktan silinmiş; geri dönülemez. Bu satırın
              var olması, listenin bir zamanlar sessizce yalan söylediği durumun ta kendisi.
            * `unmanaged` — havuzda var, kayıtta yok. Kabuktan alınmış.
            * `unknown` — ajana ulaşılamadı, karşılaştırma yapılamadı. Kayıtlı satırları `missing`
              göstermek, ajanı bir dakikalığına düşmüş bir kutuda bütün yedekleri silinmiş gibi
              göstermek olurdu.
        usedBytes:
          type: integer
          format: int64
          description: |
            Bu görüntüyü silmenin GERÇEKTEN boşaltacağı yer, havuzdan okunmuş. Yalnız havuzda
            bulunan satırlarda var.

            İçerdiği veri DEĞİL: ZFS bir görüntüye yalnız artık canlı dosya sisteminden ya da daha
            yeni bir görüntüden referans verilmeyen blokları yazıyor, yani hiçbir şeyin değişmediği
            bir terabaytlık görüntü neredeyse sıfır tutuyor. Kaydedilmiyor da: rakam zamanla
            değişiyor, ve alındığı andaki bir sayı bir hafta sonra yanlış olurdu.
    SnapshotPage:
      type: object
      required: [items, complete]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Snapshot' }
        complete:
          type: boolean
          description: |
            Liste havuzla KARŞILAŞTIRILDI mı.

            Bu alan bir zamanlar "her zaman false" diyordu, çünkü ajanda anlık görüntüleri
            listeleyecek bir işlem yoktu ve liste DEPSIS'in kendi kaydından ibaretti. Artık
            `list_snapshots` var: true olduğunda her satırın `state`'i havuza sorularak
            doldurulmuş demek, ve havuzda olup kayıtta olmayanlar da listede.

            false, yalnız ajana ulaşılamadığında. O durumda her `state` `unknown`, ve istemcinin
            bunu söylemesi gerekiyor — sessizce eski davranışa dönmek, doğrulanmamış bir listeyi
            doğrulanmış gibi göstermek olurdu.
    CreateSnapshotRequest:
      type: object
      required: [dataset]
      properties:
        dataset: { type: string, minLength: 1, maxLength: 255 }
        name:
          type: string
          maxLength: 63
          pattern: '^[A-Za-z0-9_][A-Za-z0-9._-]{0,62}$'
          description: Yoksa sunucu zaman damgasından üretir.
...
    Share:
      type: object
      required: [id, name, readOnly, published, uncPath]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        readOnly: { type: boolean }
        published:
          type: boolean
          description: |
            Samba bu paylaşımı GERÇEKTEN sunuyor mu. Veritabanında bir satır olması yetmez —
            ikisini aynı şeymiş gibi göstermek, kullanıcıyı çalışmayan bir adresi denemeye
            göndermektir.
        uncPath:
          type: string
          description: Windows'un Gezgin'e yazacağı adres, örn. `\\depsis\belgeler`.
        dataset: { type: string }
    SharePage:
      type: object
      required: [items, smbAvailable]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Share' }
        smbAvailable:
          type: boolean
          description: |
            Ajana ulaşılabiliyor ve Samba kurulu mu. false ise `published` alanlarının hepsi
            false'tur ve bu bir arıza değil, kurulmamış bir bileşendir.
    ShareCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 63
          # Ajanın `SafeComponent`'i ile birebir aynı, ve migration 0008'in `shares_name_format`
          # kısıtıyla da. Üçü ayrı yerde ama tek bir kural: bu tablonun kabul edip ajanın
          # reddettiği bir ad, veritabanında var olup diskte olmayan bir paylaşımdır.
          pattern: '^[A-Za-z0-9_][A-Za-z0-9._-]{0,62}$'
          description: |
            Paylaşımın adı, ve aynı zamanda SMB adresindeki son bileşen. Büyük/küçük harf ve
            Türkçe i ailesi katlanarak karşılaştırılır: `Belgeler` varken `belgeler` açılamaz,
            çünkü SMB istemcileri ikisini ayırt edemez.
        readOnly:
          type: boolean
          default: false
        quotaBytes:
          type: [integer, 'null']
          format: int64
          minimum: 1
          description: |
            `refquota`. Snapshot'ları HARİÇ tutar, yani yönetici anlık görüntü politikası bir
            kullanıcıyı kendi alanının dışına kilitleyemez (ADR-0008). null: kota yok.
        grants:
          type: array
          description: |
            Paylaşımın kökündeki başlangıç izinleri. Verilmezse paylaşımı açan yöneticiye tam
            izin yazılır — boş bir dizi DEĞİL: grant'sız bir paylaşım, kaldırılan istisnayı geri
            getirir, ve `422` ile reddedilir.
          minItems: 1
          items: { $ref: '#/components/schemas/Grant' }
    ShareCreated:
      type: object
      required: [share, applyingJobId]
      properties:
        share: { $ref: '#/components/schemas/Share' }
        applyingJobId:
          type: [string, 'null']
          format: uuid
          description: |
            POSIX ACL'leri yazacak işin kimliği. null ise ajana ulaşılamadı ve dosya sistemi
            henüz bu paylaşımın izinlerini bilmiyor — arayüz bunu göstermek zorunda, çünkü
            web'in kapalı dediği bir klasörün SMB'den açılması §6.2'nin adıyla yasakladığı şey.
...
    FolderPermission:
      type: string
      description: |
        §6.2'nin saydığı on bir izin, birebir. `manage` (izin yönet) ayrı: bir klasöre yazabilen
        herkesin o klasörün iznini de değiştirebilmesi, yetki modelinin kendi kendini geçersiz
        kılmasıdır.
      enum: [list, read, download, create, modify, move, delete, share, manage, versions, audit]
...
    FolderPermissions:
      type: object
      required: [effective, grants, canManage]
      properties:
        effective:
          type: array
          description: |
            ÇAĞIRANIN bu klasörde gerçekten yapabildikleri, miras hesaplandıktan sonra. Arayüz
            düğmeleri buna göre çizer. Organizasyon yöneticisi için her zaman tam küme.
          items: { $ref: '#/components/schemas/FolderPermission' }
        inheritedFrom:
          type: [string, 'null']
          format: uuid
          description: |
            `effective` kümesinin geldiği düğüm. Bu düğümün kendisiyse kendi kimliği; hiç grant
            yoksa null. "Bu izin nereden geliyor" cevaplanabilir bir soru olmalı, yoksa bir
            yöneticinin onu nereden kaldıracağını bulması imkânsız.
        grants:
          type: array
          description: |
            Bu düğümde AÇIKÇA yazılmış satırlar. `canManage` false ise BOŞ döner — kimin nereye
            erişebildiği kendi başına bir bilgi ve onu görmek de bir yetki.
          items: { $ref: '#/components/schemas/Grant' }
        canManage:
          type: boolean
          description: Çağıranın `manage` izni var mı. Arayüzün izin sekmesini göstermesi buna bağlı.
    SetFolderPermissionsRequest:
      type: object
      required: [grants]
      properties:
        grants:
          type: array
          description: |
            Bu düğümdeki grant'ların TAMAMI. Boş dizi, bu düğümdeki bütün açık izinleri kaldırır
            ve mirası yeniden yürürlüğe sokar — silmenin yolu bu, çünkü boş bir izin kümesi
            taşıyan bir grant satırı veritabanı tarafından reddediliyor.
          items: { $ref: '#/components/schemas/Grant' }
...
    PermissionWriteResult:
      type: object
      required: [applied, impact]
      properties:
        applied:
          type: boolean
          description: false ise dryRun'dı ve hiçbir şey yazılmadı.
        impact: { $ref: '#/components/schemas/PermissionImpact' }
        applyingJobId:
          type: [string, 'null']
          format: uuid
          description: |
            POSIX ACL'leri yeniden yazan iş. İzinler dosya sistemine ANINDA yansımıyor ve
            tamamlanana kadar iki katman ayrışık (ADR-0021'in kabul edilen borcu). Arayüz bunu
            "izinler uygulanıyor" diye göstermeli; sessizce tamamlanmış gibi davranmak, SMB'den
            hâlâ erişilebilen bir klasörü kapalı sanmaktır.

            Ajan ulaşılamazsa null olur ve yazma yine de yapılır — uygulama katmanı doğru,
            dosya sistemi geride. Bu durumun görünür olması, gizlenmesinden iyidir.